// web-client/src/data/api/listings/models/ListingPrivateDetails.ts
export interface ListingPrivateDetails {
    /**
     * 
     * @type {string}
     * @memberof ListingPrivateDetails
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ListingPrivateDetails
     */
    title: string;
    /**
     * 
     * @type {ListingStatus}
     * @memberof ListingPrivateDetails
     */
    status: ListingStatus;
    /**
     * 
     * @type {string}
     * @memberof ListingPrivateDetails
     */
    description?: string;
    /**
     * 
     * @type {GuestsCapacity}
     * @memberof ListingPrivateDetails
     */
    guestCapacity: GuestsCapacity;
    /**
     * 
     * @type {PropertyType}
     * @memberof ListingPrivateDetails
     */
    propertyType?: PropertyType;
    /**
     * 
     * @type {RoomType}
     * @memberof ListingPrivateDetails
     */
    roomType?: RoomType;
    /**
     * 
     * @type {Address}
     * @memberof ListingPrivateDetails
     */
    address?: Address;
    /**
     * 
     * @type {object}
     * @memberof ListingPrivateDetails
     */
    location?: object;
    /**
     * 
     * @type {Array<ListingPhotoPath>}
     * @memberof ListingPrivateDetails
     */
    photos?: Array<ListingPhotoPath>;
}